Finally added SVD support + interactive SVD test on free-use datasets

+++ src/main/java/com/kitsoft/freetify/algo/Maths.java
@@ -2,17 +2,17 @@ package com.kitsoft.freetify.algo;
 
 public final class Maths {
 
-    public static double cosineSimilarity(int[] a, int[] b) {
+    public static double cosineSimilarity(double[] a, double[] b) {
         return dotProduct(a, b) / Math.sqrt(magnitudeSq(a) * magnitudeSq(b));
     }
 
-    public static int dotProduct(int[] a, int[] b) {
-        int sum = 0;
+    public static double dotProduct(double[] a, double[] b) {
+        double sum = 0;
         for (int i = 0; i < a.length; i++) sum += a[i] * b[i];
         return sum;
     }
 
-    public static int magnitudeSq(int[] a) {
+    public static double magnitudeSq(double[] a) {
         return dotProduct(a, a);
     }
 }
